mcp_testing_server/requirements.txt
mcp>=1.0.0
patchright
pylint
radon
pyautogui
pillow
